Pages locked down for admin use only for CRUD

+++ src/main/java/com/uball/uballapp/controller/MachineController.java
@@ -17,10 +17,9 @@ public class MachineController {
         this.machineDao = machineDao;
     }
 
-
     /**
      * ounces URI accessed, displays machines page
-     * */
+     */
     @GetMapping("/createmachine")
     public String index(Model model) {
 
@@ -37,57 +36,93 @@ public class MachineController {
         }
 
     }
+
     /**
      * returns create machine form
-     * */
+     */
     @GetMapping("/machines/create")
     public String create(Model model) {
-        model.addAttribute("machinesForm", new Machine());
-        return "machines/create";
+
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            model.addAttribute("machinesForm", new Machine());
+            return "machines/create";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
+
     /**
      * accessed through create machine form and inserts into DB before redirecting to machines index page
-     * */
+     */
     @PostMapping("/machines/create")
     public String insert(@ModelAttribute Machine machine) {
-        machineDao.save(machine);
-        return "redirect:/createmachine";
+
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            machineDao.save(machine);
+            return "redirect:/createmachine";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
 
     /**
      * when URI accessed from machines index page, displays edit machine page by id
-     * */
+     */
     @GetMapping("/machines/{id}/edit")
     public String edit(@PathVariable long id, Model model) {
-        model.addAttribute("thisMachine", machineDao.findOne(id));
-        return "machines/edit";
+
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            model.addAttribute("thisMachine", machineDao.findOne(id));
+            return "machines/edit";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
 
     /**
      * accessed through edit machine form and inserts into DB before redirecting to machines index page
-     * */
+     */
     @PostMapping("/machines/{id}/edit")
     public String update(@PathVariable long id, @ModelAttribute Machine machine) {
-        machineDao.save(machine);
-        return "redirect:/createmachine";
+
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            machineDao.save(machine);
+            return "redirect:/createmachine";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
 
     @GetMapping("/machines/{id}/delete")
     public String deleteMachine(@PathVariable long id, @ModelAttribute Machine machine) {
-        machineDao.delete(id);
-        return "redirect:/createmachine";
+
+        User userSession = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        long userId = userSession.getId();
+        boolean isAdmin = userSession.isAdmin();
+
+        if (isAdmin) {
+            machineDao.delete(id);
+            return "redirect:/createmachine";
+        } else {
+            return "redirect:/userprofile";
+        }
     }
-
-
-
-
-
-
-
-
-
-
-
 
 
 }
